[feat] combine parse_pdfs and load_documents
