+++ etl/textrank/settings/config.py
@@ -1,9 +1,10 @@
 import os
 from nltk.stem.snowball import EnglishStemmer
+from pkgutil import get_data
 
 # Variables
 # ---------------------------------------------------------------------------------------------
-stopwordsFile = 'nltk.txt' # file to grab stopwords from (in stopwords folder)
+stopwords = {word.strip() for word in str(get_data('data', 'stopwords.txt').decode('utf-8')).split('\n')}
 stemmer = EnglishStemmer() # stemmer class
 wordPattern = "^[^\W\d_]+$" # regex pattern to match a word
 epsilon = 1e-4 # epsilon value for algorithm
@@ -16,10 +17,10 @@ delta = 1e-7 # delta value for algorithm
 # ---------------------------------------------------------------------------------------------
 
 # Create stopwords
-with open('{0}\\stopwords\\{1}'.format(os.path.dirname(os.path.realpath(__file__)), stopwordsFile), 'r') as f:
-    words = ()
-    for word in f.readlines():
-        words += (word.rstrip(),)
-    stopwords = frozenset(words)
+# with open('{0}\\stopwords\\{1}'.format(os.path.dirname(os.path.realpath(__file__)), stopwordsFile), 'r') as f:
+#     words = ()
+#     for word in f.readlines():
+#         words += (word.rstrip(),)
+#     stopwords = frozenset(words)
 
 # ---------------------------------------------------------------------------------------------
